Delete running GMainLoop in player initialization; rename FakeSystemMedia to MPRISNowPlaying
