+++ Sources/MusicPlayer/Players/SystemMedia.swift
@@ -164,82 +164,4 @@ private extension Notification.Name {
     static let mediaRemoteNowPlayingApplicationPlaybackStateDidChange = Notification.Name("kMRMediaRemoteNowPlayingApplicationPlaybackStateDidChangeNotification")
 }
 
-#elseif os(Linux)
-
-import playerctl
-
-extension MusicPlayers {
-    
-    public final class FakeSystemMedia: NowPlaying {
-        
-        private let manager: UnsafeMutablePointer<PlayerctlPlayerManager>
-        private var signals: [gulong] = []
-        
-        public init?() {
-            guard let manager = playerctl_player_manager_new(nil) else {
-                return nil
-            }
-            self.manager = manager
-            
-            GEventLoop.start()
-            
-            var players: [MusicPlayerProtocol] = []
-            var playerNames: UnsafeMutablePointer<GList>? = playerctl_list_players(nil)
-            while (playerNames != nil) {
-                let playerName = playerNames!.pointee.data.assumingMemoryBound(to: PlayerctlPlayerName.self)
-                let player = playerctl_player_new_from_name(playerName, nil)
-                if player == nil {
-                    continue
-                }
-                playerctl_player_manager_manage_player(manager, player)
-                players.append(MPRIS(player: player!))
-                playerNames = playerNames!.pointee.next
-            }
-            
-            super.init(players: players)
-            
-            let onNameAppeared: @convention(c) (UnsafeMutablePointer<PlayerctlPlayerManager>?,
-                                                UnsafeMutablePointer<PlayerctlPlayerName>?,
-                                                UnsafeMutableRawPointer?) -> Void
-                = { manager, name, data in
-                    let `self`: FakeSystemMedia = Unmanaged.fromOpaque(data!).takeUnretainedValue()
-                    if let player = playerctl_player_new_from_name(name, nil) {
-                        playerctl_player_manager_manage_player(`self`.manager, player)
-                        `self`.players.append(MPRIS(player: player))
-                    }
-                }
-            
-            let onPlayerVanished: @convention(c) (UnsafeMutablePointer<PlayerctlPlayerManager>?,
-                                                  UnsafeMutablePointer<PlayerctlPlayer>?,
-                                                  UnsafeMutableRawPointer?) -> Void
-                = { manager, player, data in
-                    if player == nil {
-                        return
-                    }
-                    data?.unretainedCast(to: FakeSystemMedia.self).players.removeAll { ($0 as? MPRIS)?.player == player }
-                }
-            
-            let pself = Unmanaged.passUnretained(self).toOpaque()
-            signals.append(
-                g_signal_connect_data(manager, "name-appeared", unsafeBitCast(onNameAppeared, to: GCallback?.self), pself, nil, G_CONNECT_AFTER)
-            )
-            signals.append(
-                g_signal_connect_data(manager, "player-vanished", unsafeBitCast(onPlayerVanished, to: GCallback?.self), pself, nil, G_CONNECT_AFTER)
-            )
-        }
-        
-        deinit {
-            for var signal in signals {
-                g_clear_signal_handler(&signal, manager)
-            }
-            g_free(manager)
-        }
-    }
-}
-
-extension MusicPlayers {
-    
-    public typealias SystemMedia = FakeSystemMedia
-}
-
 #endif
