+++ Sources/MusicPlayer/Utilities/Typealias.swift
@@ -27,7 +27,6 @@ public typealias Image = UIImage
 
 #else
 
-// TODO: image on linux?
 public typealias Image = URL
 
 #endif
